feat(contracts): adjusted namespaces #86
