+++ Attest.Testing.Contracts/Configuration/IConfigurationSectionKeySplitter.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 
-// ReSharper disable once CheckNamespace
 namespace Attest.Testing.Configuration
 {
     public interface IConfigurationSectionKeySplitter
