+++ Attest.Testing.Contracts/Lifecycle/IStopApplicationService.cs
@@ -1,4 +1,4 @@
-﻿namespace Attest.Testing.Contracts
+﻿namespace Attest.Testing.Lifecycle
 {
     /// <summary>
     /// Represents means of stopping the running application.
