+++ Attest.Testing.Core.Specs/Serialization/ISimpleProvider.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+// ReSharper disable once CheckNamespace
+namespace Attest.Testing.Serialization.Specs
+{
+    public interface ISimpleProvider
+    {
+        IEnumerable<SimpleItemDto> GetSimpleItems();
+    }
+}
