+++ Attest.Testing.Core.Specs/Serialization/SerializationScenarioDataStore.cs
@@ -1,9 +1,10 @@
 ﻿using Attest.Fake.Data;
 using Attest.Testing.Context.SpecFlow;
-using Attest.Testing.Contracts;
+using Attest.Testing.FakeData;
 using TechTalk.SpecFlow;
 
-namespace Attest.Testing.Core.Specs
+// ReSharper disable once CheckNamespace
+namespace Attest.Testing.Serialization.Specs
 {
     internal sealed class SerializationScenarioDataStore : ScenarioDataStoreBase
     {
