+++ Attest.Testing.Core.Specs/Serialization/SimpleItemDto.cs
@@ -1,0 +1,10 @@
+﻿// ReSharper disable once CheckNamespace
+namespace Attest.Testing.Serialization.Specs
+{
+    public class SimpleItemDto
+    {
+        public string Name { get; set; }
+        public double Price { get; set; }
+        public int Quantity { get; set; }
+    }
+}
