+++ Attest.Testing.Core/EndToEnd/EndToEndScenarioDataStore.cs
@@ -1,5 +1,6 @@
 using Attest.Testing.Context;
-using Attest.Testing.Contracts;
+using Attest.Testing.FakeData;
+using Attest.Testing.Lifecycle;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.EndToEnd
