+++ Attest.Testing.Core/EndToEnd/StartApplicationService.cs
@@ -1,5 +1,6 @@
 using Attest.Fake.Data;
-using Attest.Testing.Contracts;
+using Attest.Testing.Application;
+using Attest.Testing.Lifecycle;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.EndToEnd
