+++ Attest.Testing.Core/EndToEnd/StartLocalApplicationService.cs
@@ -1,6 +1,7 @@
 using System.IO;
-using Attest.Testing.Contracts;
+using Attest.Testing.Application;
 using Attest.Testing.Core;
+using Attest.Testing.Lifecycle;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.EndToEnd
