+++ Attest.Testing.Core/FakeData/RegistratorExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using Attest.Fake.Data;
-using Attest.Testing.Contracts;
 using Solid.Practices.IoC;
 
 // ReSharper disable once CheckNamespace
