+++ Attest.Testing.Core/Integration/StartLocalApplicationService.cs
@@ -1,4 +1,4 @@
-using Attest.Testing.Contracts;
+using Attest.Testing.Lifecycle;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.Integration
