+++ Attest.Testing.Core/Lifecycle/ResolverExtensions.cs
@@ -1,4 +1,4 @@
-﻿using Attest.Testing.Contracts;
+﻿using Attest.Testing.Modularity;
 using Solid.Practices.IoC;
 
 // ReSharper disable once CheckNamespace
