+++ Attest.Testing.Core/Lifecycle/SetupServiceBase.cs
@@ -1,4 +1,4 @@
-using Attest.Testing.Contracts;
+
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.Lifecycle
