+++ Attest.Testing.Core/Lifecycle/StartDynamicApplicationModuleService.cs
@@ -1,11 +1,10 @@
 using System.IO;
-using Attest.Testing.Contracts;
 using Attest.Testing.Core;
+using Attest.Testing.Modularity;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.Lifecycle
 {
-    /// <inheritdoc />
     public sealed class StartDynamicApplicationModuleService : IStartDynamicApplicationModuleService
     {
         /// <inheritdoc />
