+++ Attest.Testing.Core/Lifecycle/StaticSetupService.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
-using Attest.Testing.Contracts;
+using Attest.Testing.Modularity;
 using Solid.Core;
 using Solid.Practices.IoC;
