+++ Attest.Testing.Core/Lifecycle/TeardownService.cs
@@ -1,4 +1,4 @@
-using Attest.Testing.Contracts;
+using Attest.Testing.Application;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.Lifecycle
