+++ Attest.Testing.Core/Management/BootstrapperExtensions.cs
@@ -1,5 +1,4 @@
-﻿using Attest.Testing.Contracts;
-using Solid.Bootstrapping;
+﻿using Solid.Bootstrapping;
 using Solid.Extensibility;
 
 // ReSharper disable once CheckNamespace
