+++ Attest.Testing.Core/Management/ProcessManagementServiceBase.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using Attest.Testing.Contracts;
 
 // ReSharper disable once CheckNamespace
 namespace Attest.Testing.Management
