+++ Attest.Testing.Core/Management/RegistratorExtensions.cs
@@ -1,4 +1,3 @@
-using Attest.Testing.Contracts;
 using Solid.Practices.IoC;
 
 // ReSharper disable once CheckNamespace
